Moved api page to components
